Revert "Version 0.2.0 - Improved searching (#5)"
This reverts commit 31aa357405290f7542f65a4c31326eccfd14b9e4.
