+++ .gitignore
@@ -14,11 +14,3 @@ data/
 input/
 output/
 script_launcher.bat
-
-# Hidden docx files
-*~*.docx
-
-# Pyinstaler
-dist/
-build/
-*.spec
