+++ requirements.txt
@@ -1,4 +1,4 @@
 lxml==4.5.2
-pip==20.2.3
+pip==20.2.2
 python-docx==0.8.10
 setuptools==50.3.0
