+++ src/docxHandlers/docxReader.py
@@ -1,110 +1,34 @@
 import re
 import docx
-import lxml
 from src.common.defines import *
 from src.common import pathHelpers
 from src.cmdInterface import ansiColorHelper, cmdProgressBar
 
 
-def extract_acro_word(filepath, acro_dict_handler):
-    """Main function. Opens a docx file and extracts its acronyms
-
-    :param filepath: Path string to a docx file
-    :param acro_dict_handler: Acronym dictionary objects
-    """
-    # 1. Open file
-    doc_obj = docx.Document(filepath)
-    acro_dict_handler.str_file = pathHelpers.get_filename_from_path(filepath)
-
-    # 2. Search for acronyms and extract them
-    print("Extrayendo acrónimos del documento")
-    extract_acro_from_paragraphs(doc_obj, acro_dict_handler)
-    extract_acro_from_tables(doc_obj, acro_dict_handler)
-    extract_acro_from_sections(doc_obj, acro_dict_handler)
-
-    # 3. Perform second search to find acronyms or abbreviates that do not match with the regex and confirm the ones in
-    # the document acronym table that have not been found yet (Ej: ExCOMMS, JdP)
-    second_regex = "" # All words combined into a regex. It speeds up drastically the search as only one pass is needed
-
-    if config_use_acro_from_doc_table:  # todo, revisar acceso directo a dict handler
-        for acro_key in acro_dict_handler.acros_doc_table.keys():
-            if acro_key not in acro_dict_handler.acros_found:
-                second_regex += re.escape(acro_key) + r'|'
-
-    if config_use_non_matching_acro_from_db:
-        for acro_key in acro_dict_handler.obj_db.list_no_regex:
-            if acro_key not in acro_dict_handler.acros_found and acro_key not in second_regex:
-                second_regex += re.escape(acro_key) + r'|'
-
-    if second_regex != "":
-        print("Repasando la búsqueda con acrónimos especiales")
-        # Add word boundries to regex
-        second_regex = r'\b(' + second_regex[:-1] + r')(?![\wÁÉÍÓÚ])'
-        # The last part fixes abbreviates. \b does not match '. ' as both are non alphanumeric
-        # Alternate: r'\b('+second_regex[:-1]+r')(\b|(?=\W))'
-        extract_acro_from_paragraphs(doc_obj, acro_dict_handler, second_regex)
-        extract_acro_from_tables(doc_obj, acro_dict_handler, second_regex)
-        extract_acro_from_sections(doc_obj, acro_dict_handler, second_regex)
-
-
-def accepted_text(docx_elem, docx_elem_xml, doc_nsmap):
-    str_accepted_text = ""
-    if "w:ins" in docx_elem_xml:
-        # print(docx_elem_xml)
-        target = lxml.etree.XML(docx_elem_xml)
-        # Search all paragraphs in the XML
-        for paragraph in target.xpath('//w:p', namespaces=doc_nsmap):
-            # Search all runs and inserted runs (Tack Changes). Deleted runs (w:del) are skipped
-            for text_run in paragraph.xpath('w:r | w:ins/w:r', namespaces=doc_nsmap):
-                # Handle linebreaks first. It seems that appear in their own run or before text
-                if text_run.xpath('w:br', namespaces=doc_nsmap):
-                    str_accepted_text += '\n'
-                for text_tag in text_run.xpath('w:t', namespaces=doc_nsmap):
-                    str_accepted_text += text_tag.text
-            str_accepted_text += "\n"  # Append new line between paragraphs (There can be multiple pgphs in tables)
-        str_accepted_text = str_accepted_text[:-1]  # Remove last paragraph new line char
-    else:
-        # Use python docx when possible as it has the text already extracted
-        str_accepted_text = docx_elem.text
-    return str_accepted_text
-
-
-def extract_acro_from_str(str_in_raw, acro_dict_handler, regex_in=""):
+# Fixme: Move dict responsabilities to "acroDictHandler"
+def extract_acro_from_str(str_in, dict_found_acro, regex_in=""):
     """Finds acronyms in a text string and stores them into a dictionary
 
     :param str_in: Input text string
-    :param acro_dict_handler: Acronym dictionary objects
+    :param dict_found_acro: Acronyms found dictionary
     :param regex_in: Regex string
     """
     # 1. Find Acronyms as regex matches of groups of N Capital Leters. The regex can be changed for other uses
     if regex_in == "":
         regex_in = define_regex_acro_find
 
-    # Line breaks are removed to reduce space used when outputting the context string to console
-    str_in = str_in_raw.replace('\n', config_new_line_separator)
     re_results = re.finditer(regex_in, str_in)
 
     if re_results:
         # 2. Iterate trough all matches (There could be multiple acronyms in a paragraph)
         for re_result in re_results:
-            if re_result.group(0) == "AEO":
-                a=10
-            # 3. Save context for user analysis (Only part of the string)
+            # 3. Save context for user analysis (Only some of the paragraph)
             context_width = 200
-            # Ensure the context width is achieved
-            prev_context = context_width / 2
-            next_context = context_width / 2
-            # Check if there are unused characters at front/back after adding/subtracting the context. If so, add them
-            # to the opposite side.
-            char_diff_ini = re_result.start(0) - context_width / 2
-            if char_diff_ini < 0:
-                next_context -= char_diff_ini  # Addition, double negative
-            char_diff_end = len(str_in) - (re_result.start(0) + context_width / 2)
-            if char_diff_end < 0:
-                prev_context -= char_diff_end  # Addition, double negative
-
-            context_substr_ini = int(max(re_result.start(0) - prev_context, 0))
-            context_substr_end = int(min(re_result.start(0) + next_context, len(str_in)))
+            # Ensure the width is achieved by checking the available width
+            # Fixme: Solo recorta bien el contexto si el acro está al principio
+            context_substr_ini = int(max(re_result.start(0) - context_width / 2, 0))
+            context_substr_end = int(min(re_result.start(0) +
+                                         (context_width - (re_result.start(0) - context_substr_ini)), len(str_in)))
 
             # Context with acronym remarked using ANSI Color Codes
             str_context = ansiColorHelper.highlight_substr(
@@ -113,30 +37,36 @@ def extract_acro_from_str(str_in_raw, acro_dict_handler, regex_in=""):
                 ansiColorHelper.AnsiColorCode.CYAN)
 
             # 4. Create empty dict if acronym is first found
-            acro_dict_handler.add_acronym_found(re_result.group(0), str_context)
+            if re_result.group(0) not in dict_found_acro:
+                dict_found_acro[re_result.group(0)] = {'Count': 0, 'Context': []}
+
+            # 5. Store data of acronym in dict
+            dict_found_acro[re_result.group(0)]['Count'] += 1
+            dict_found_acro[re_result.group(0)]['Context'].append(str_context)
 
 
-def extract_acro_from_paragraphs(doc_obj, acro_dict_handler, regex_in=""):
+def extract_acro_from_paragraphs(doc_obj, dict_found_acro, regex_in=""):
     """ Iterates through all paragraph blocks of the document and stores all acronyms found
 
     :param doc_obj: Python-docx object
-    :param acro_dict_handler: Acronym dictionary objects
+    :param dict_found_acro: Acronyms found dictionary
     :param regex_in: Regex string
     """
     # Iterate trough all paragraphs
     obj_progress_bar = cmdProgressBar.CmdProgressBar(len(doc_obj.paragraphs), "Párrafos")
-    for i, paragraph in enumerate(doc_obj.paragraphs):
-        str_accepted_text = accepted_text(paragraph, paragraph._p.xml, doc_obj.element.nsmap)
-        extract_acro_from_str(str_accepted_text, acro_dict_handler, regex_in=regex_in)
+    for i in range(len(doc_obj.paragraphs)):
+        par_txt = doc_obj.paragraphs[i].text
+        extract_acro_from_str(par_txt, dict_found_acro, regex_in=regex_in)
 
         obj_progress_bar.update(i + 1)
 
 
-def extract_acro_from_tables(doc_obj, acro_dict_handler, regex_in=""):
+def extract_acro_from_tables(doc_obj, dict_found_acro, dict_acro_table, regex_in=""):
     """ Iterates through all document tables and stores all acronyms found. Processes the acro-table if found
 
     :param doc_obj: Python-docx object
-    :param acro_dict_handler: Acronym dictionary objects
+    :param dict_found_acro: Acronyms found dictionary
+    :param dict_acro_table: Acronyms table document dictionary
     :param regex_in: Regex string
     """
     # 1. Iterate trough table objects
@@ -146,17 +76,19 @@ def extract_acro_from_tables(doc_obj, acro_dict_handler, regex_in=""):
         for j, row in enumerate(table.rows):
             row_cell_list = []
             for cell in row.cells:
-                row_cell_list.append(accepted_text(cell, cell._tc.xml, doc_obj.element.nsmap))
+                row_cell_list.append(cell.text)
             row_text = config_tb_col_separator.join(row_cell_list)
 
             if j == 0:
                 if is_acronym_table_header(row_text):
-                    process_acro_table(table, acro_dict_handler)
+                    process_acro_table(table, dict_acro_table)
                     break  # Do not process acronym table as found acronyms
 
-            extract_acro_from_str(row_text, acro_dict_handler, regex_in=regex_in)
+            # Line breaks are removed to reduce space used when outputting the context string to console
+            extract_acro_from_str(row_text.replace('\n', config_new_line_separator), dict_found_acro, regex_in=regex_in)
 
         obj_progress_bar.update(i + 1)
+
 
 def is_acronym_table_header(row_input):
     """Checks if the string matches any document acronyms table header defined
@@ -172,19 +104,20 @@ def is_acronym_table_header(row_input):
     return flag_return
 
 
-def process_acro_table(acro_table, acro_dict_handler):
+def process_acro_table(acro_table, dict_acro_table):
     """Processes the document acronym table to extract already defined acronyms. Currently the format is set as:
     Acronym | Definitions as "original text (Translated)" separated by line breaks
 
     :param acro_table: Python-docx table object
-    :param acro_dict_handler: Acronyms table document dictionary
+    :param dict_acro_table: Acronyms table document dictionary
     """
     # This function only works if the following format is used:
     # Acronym | Definitions as "original text (Translated)" separated by line breaks
     # When a table is found with a different number of rows it is skiped
     # Todo: Allow for custom tables via configure file
     # Fixme: Prevent explosion if table has merged lines
-    if not acro_dict_handler.flag_doc_table_processed and len(acro_table.rows[0].cells) == 2:
+    if 'Flag_Processed' not in dict_acro_table and len(acro_table.rows[0].cells) == 2:
+        print(" (Procesando tabla de acrónimos del documento)", end="")
         for i, row in enumerate(acro_table.rows):
             if i == 0:  # Skip header
                 continue
@@ -196,54 +129,66 @@ def process_acro_table(acro_table, acro_dict_handler):
             if acronym == "":
                 continue
 
-            # Process definitions
-            for raw_def in definitions:
-                main_def = raw_def.strip()
-                trans_def = ""
+            # Create dict entry
+            if acronym not in dict_acro_table:
+                # Some tables have duplicated lines for multiple definitions
+                dict_acro_table[acronym] = {'Def': []}
 
-                if '(' in main_def:
-                    re_match = re.fullmatch("(.*)\((.*)\)", main_def)  # Todo: Fix "Def (Expl) (Def_es (Expl_es))"
+            # Process definitions
+            for j in range(len(definitions)):
+                aux_dict_def = dict()
+
+                definition = definitions[j].strip()
+                if '(' not in definition:
+                    main_def = definition
+                else:
+                    main_def = ""
+                    re_match = re.fullmatch("(.*)\((.*)\)", definition)  # Todo: Fix "Def (Expl) (Def_es (Expl_es))"
+
                     if re_match:
                         main_def = re_match.group(1).strip()
-                        trans_def = re_match.group(2).strip()
+                        aux_dict_def['Translation'] = re_match.group(2).strip()
 
-                acro_dict_handler.add_acronym_doc_table(acronym, main_def, trans_def)
-    acro_dict_handler.flag_doc_table_processed = True
+                aux_dict_def['Main'] = main_def
+                dict_acro_table[acronym]['Def'].append(aux_dict_def)
+        dict_acro_table['Flag_Processed'] = True
 
 
-def extract_acro_from_sections(doc_obj, acro_dict_handler, regex_in=""):
-    """ Iterates through all sections blocks of the document and stores all acronyms found
+def extract_acro_word(filepath, acro_dict_handler):
+    """Main function. Opens a docx file and extracts its acronyms
 
-    :param doc_obj: Python-docx object
+    :param filepath: Path string to a docx file
     :param acro_dict_handler: Acronym dictionary objects
-    :param regex_in: Regex string
     """
-    obj_progress_bar = cmdProgressBar.CmdProgressBar(len(doc_obj.sections), "Secciones")
-    nsmap = doc_obj.element.nsmap
-    for i, section in enumerate(doc_obj.sections):
-        # Sections include headers and footers
-        extract_acro_from_header(section.header, acro_dict_handler, nsmap, regex_in)
-        extract_acro_from_header(section.footer, acro_dict_handler, nsmap, regex_in)
+    # 1. Open file
+    doc_obj = docx.Document(filepath)
+    acro_dict_handler.str_file = pathHelpers.get_filename_from_path(filepath)
 
-        obj_progress_bar.update(i + 1)
+    # 2. Search for acronyms and extract them
+    print("Extrayendo acrónimos del documento")
+    extract_acro_from_paragraphs(doc_obj, acro_dict_handler.acros_found)
+    extract_acro_from_tables(doc_obj, acro_dict_handler.acros_found, acro_dict_handler.acros_doc_table)
 
+    # 3. Perform second search to find acronyms or abbreviates that do not match with the regex and confirm the ones in
+    # the document acronym table that have not been found yet (Ej: ExCOMMS, JdP)
+    second_regex = "" # All words combined into a regex. It speeds up drastically the search as only one pass is needed
 
-def extract_acro_from_header(docx_section_obj, acro_dict_handler, nsmap, regex_in=""):
-    """ Processes all paragraphs and tables of a header or footer object
+    if config_use_acro_from_doc_table:
+        for acro_key in acro_dict_handler.acros_doc_table.keys():
+            if acro_key not in acro_dict_handler.acros_found:
+                second_regex += re.escape(acro_key) + r'|'
 
-    :param docx_section_obj: Python-docx object (section.header or section.footer)
-    :param acro_dict_handler: Acronym dictionary objects
-    :param regex_in: Regex string
-    """
-    for paragraph in docx_section_obj.paragraphs:
-        str_accepted_text = accepted_text(paragraph, paragraph._p.xml, nsmap)
-        extract_acro_from_str(str_accepted_text, acro_dict_handler, regex_in=regex_in)
+    if config_use_non_matching_acro_from_db:
+        for acro_key in acro_dict_handler.list_no_regex:
+            if acro_key not in acro_dict_handler.acros_found and acro_key not in second_regex:
+                second_regex += re.escape(acro_key) + r'|'
 
-    for table in docx_section_obj.tables:
-        for j, row in enumerate(table.rows):
-            row_cell_list = []
-            for cell in row.cells:
-                row_cell_list.append(accepted_text(cell, cell._tc.xml, nsmap))
-            row_text = config_tb_col_separator.join(row_cell_list)
-            # Acronym table not expected in header/footers
-            extract_acro_from_str(row_text, acro_dict_handler, regex_in=regex_in)
+    if second_regex != "":
+        print("Repasando la búsqueda con acrónimos especiales")
+        # Add word boundries to regex
+        second_regex = r'\b(' + second_regex[:-1] + r')(?![\wÁÉÍÓÚ])'
+        # The last part fixes abbreviates. \b does not match '. ' as both are non alphanumeric
+        # Alternate: r'\b('+second_regex[:-1]+r')(\b|(?=\W))'
+        extract_acro_from_paragraphs(doc_obj, acro_dict_handler.acros_found, second_regex)
+        extract_acro_from_tables(doc_obj, acro_dict_handler.acros_found, acro_dict_handler.acros_doc_table,
+                                 second_regex)
